Floats

+++ test.py
@@ -1,6 +1,6 @@
 operator = input("Enter an operator (+, -, *, /): ")
-num1 = int(input("Enter first num: "))
-num2 = int(input("Enter second num: "))
+num1 = float(input("Enter first num: "))
+num2 = float(input("Enter second num: "))
 
 if (operator == "+") :
     print(f"{num1} + {num2} = {num1 + num2}")
